Moved time helper functions to it's own "time.py" file.
